MAJOR: make permutation constants variable by RNG
There is a practical and a theoretical reason for this. Theoretically
speaking, we do not know whether the constants I drew at some point are
independent enough, so it makes sense to allow the user to vary these
based on a generator. Practically speaking, we can now use the same
benchmark strategy as used by Awad et al. in BGHT, namely to vary the
permutation constants instead of the data.

This is a major refactor, but the tests pass, and in terms of core
implementation, only significant changes are made in quotient.cuh.

BasicPermute has moved to the quotient_overhead benchmark only, as I
found (with this very benchmark) that the overhead of having the
constants in variables instead of hard-coded is negligible.