benchmarks: address confusing FoP-measurements
The FoP-benchmark happily reports measurements for a before-fill-factor
of 0.75 and an after-fillfactor of 0.5. This simply corresponds to
before=after=.75 (with a different mix of duplicate keys), but it's
better to just remove them.

First shown measurement is now before=0.7 and after=.75, still slightly
confusing but other solutions require changing the benchmark code, which
I am not ready to do right now.