benchmarks: do not store results where possible
This saves memory, and bandwidth. We do this for all puts and for
iceberg fop (cannot do it for Cuckoo fop because Cuckoo uses the results
during its find-phase). We do not use it for find as BGHT does not do it
in their benchmarks --- even though I do not find any significant
performance benefit to not storing the results.