artifact: comparable results on 2080 Ti, 3090, L40s
With the new benchmarks, they are not exactly the same, hence
"comparable". In particular, with the 3090 the compact
16-slots-per-bucket are slightly faster than compact 32-per-bucket, and
on the L40s the 16 bit compactness seems to have a slightly greater
speedup, compared to with the RTX 4090. But there was no space to expand
on this in the article, so we will also keep it curt here.